fix: use beam section in tests (#346)
* chore: rename GenericSection and GenericSectionCalculator (#341)

* fix: add last point to complete domain in nm_interaction_domain (#342)

* fix: raise a DeprecationWarning for the GenericSection (#344)

* Use BeamSection instead of GenericSection in tests
