+++ tests/test_geometry/test_profiles/test_w.py
@@ -16,7 +16,7 @@ from structuralcodes.geometry.profiles import (
 from structuralcodes.materials.basic import (
     ElasticMaterial,
 )
-from structuralcodes.sections._generic import GenericSection
+from structuralcodes.sections import BeamSection
 
 
 def load_w_profiles_data():
@@ -238,7 +238,7 @@ def test_w_massprops_polygon(
     # Create geometry
     geo = SurfaceGeometry(W.get_polygon(profile_name), material=steel)
     # Create the section
-    sec = GenericSection(geo)
+    sec = BeamSection(geo)
 
     # Get values from section
     a_section = sec.gross_properties.area
